requirements.txt
streamlit==1.25.0
tensorflow==2.16.1  # Latest version supporting Python 3.12
numpy==1.26.4  # Compatible with TensorFlow 2.16.1
pillow==9.5.0  # Compatible with Streamlit 1.25.0
matplotlib==3.8.2  # Updated version